# public_docs/swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/tasks/create": {
      "post": {
        "description": "Create Tasks",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Create Tasks",
        "operationId": "Tasks",
        "parameters": [
          {
            "name": "tasks",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#definitions/TaskCreateRequestBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/GenericResponse"
            }
          },
          "400": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "error decoding request body"
                },
                "message": {
                  "type": "string",
                  "example": "unable to process request"
                }
              }
            }
          },
          "500": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 500
                },
                "error": {
                  "type": "string",
                  "example": "Unable to process the request"
                },
                "message": {
                  "type": "string",
                  "example": "An error occurred"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/delete": {
      "delete": {
        "description": "Delete Tasks",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Delete Tasks",
        "operationId": "Tasks",
        "parameters": [
          {
            "example": "Z34b",
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/GenericResponse"
            }
          },
          "400": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "error decoding request body"
                },
                "message": {
                  "type": "string",
                  "example": "unable to process request"
                }
              }
            }
          },
          "500": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 500
                },
                "error": {
                  "type": "string",
                  "example": "Unable to process the request"
                },
                "message": {
                  "type": "string",
                  "example": "An error occurred"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/get": {
      "get": {
        "description": "Get Tasks",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Get Tasks",
        "operationId": "GetTasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "InProgress"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "int",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetTasksResponse"
            }
          },
          "400": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "status/priority sent in query string is not valid"
                },
                "message": {
                  "type": "string",
                  "example": "unable to process request"
                }
              }
            }
          },
          "500": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 500
                },
                "error": {
                  "type": "string",
                  "example": "Unable to process the request"
                },
                "message": {
                  "type": "string",
                  "example": "An error occurred"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/update": {
      "put": {
        "description": "Update Tasks",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Update Tasks",
        "operationId": "Tasks",
        "parameters": [
          {
            "name": "tasks",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#definitions/TaskCreateRequestBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/GenericResponse"
            }
          },
          "400": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "error decoding request body"
                },
                "message": {
                  "type": "string",
                  "example": "unable to process request"
                }
              }
            }
          },
          "500": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "int",
                  "example": 500
                },
                "error": {
                  "type": "string",
                  "example": "Unable to process the request"
                },
                "message": {
                  "type": "string",
                  "example": "An error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "GenericResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "task-manager/dto"
    },
    "GetTasksResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "tasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Task"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "task-manager/dto"
    },
...
    "Task": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "dead_line": {
          "type": "string",
          "x-go-name": "DeadLine"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "priority": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Priority"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "task-manager/model"
    },
...
  }
}